PDF_Reader/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
PyPDF2
faiss-cpu
google-generativeai
